requirements.txt
asgiref==3.6.0
attrs==22.2.0
autopep8==2.0.2
colorama==0.4.6
Django==4.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
iniconfig==2.0.0
packaging==23.0
pluggy==1.0.0
pycodestyle==2.10.0
PyJWT==2.6.0
pytest==7.2.2
pytz==2023.3
sqlparse==0.4.3
tzdata==2023.3
